Adding comments, creating utils file and benchmark file for each check.
