+++ Method.py
@@ -3,7 +3,5 @@ from enum import Enum
  
 class Method(Enum):
     EGH = 1
-    BINARY_COVERING_ARRAY = 2
-    RECURSIVE_ARRAY = 3
-    EXTENDED_HAMMING_CODE = 4,
-    BCH = 5
+    EXTENDED_HAMMING_CODE = 2,
+    BCH = 3
